Remove remaining tasks and previous network utility, move bitmap worker to util
